feat(logic): use standarize json for logic operators
A result_1 is considered equal to result_2 when both of their article numbers are equal. This
is the unique identifier used by IEEE. If an article number is not present for any of the
results, their "title", "publication_year", "content_type", and "publisher" have to be equal.

BREAKING CHANGE: Search results from both API and Scrap can now be mixed when performing logic
  operations.