fix: correctly add a data field to multi-word phrases
Multi-word phrases are defined as being surrounded by double quotes, and containing more than
one word inside the double quotes.
Some refactoring was done to the way we test for IEEE search operators, and how terms with
parentheses are found.
The search query, before encoding, is also shown as part of debugging