feat(scrap)!: use puppeteer browser instead of local one
Puppeteer is having problems with local browsers, it can be inconsistant.
Even though we can't package the app using puppeteer browser, this is a step gap until we can
distribute the browser while packaging. Probably by checking for it, and then downloading it.

BREAKING CHANGE: Use of puppeteer browser, so packaging the search script is not possible.