fix: run eslint and prettier (#862)
eslint-unicorn wasn't checking for errors, it needed a base config that was never there.
This commit uses the @eslint/js recommended configuration, together with node and import plugins.
In addition, Prettier is also run when committing to ensure a consisten standard.

This commit fixes a single misspelled function in cli-search.js, but mostly just fixes eslint.