feat!: add option to search all content types or just a subset
We now can choose what content we search for both scrap and API.
Sci-hub doesn't have access to books, standards, and courses.
In addition, this types are less relevant than the default ones.

BREAKING CHANGE: The default only searches for Journals, Magazines, and Conferences.
Use the --all-content-types option to search for all types.